+++ frontend/src/Components/common/utils.jsx
@@ -1,0 +1,3 @@
+export function isValidProjectName(str) {
+	return str !== "";
+}
