+++ frontend/src/Pages/Dashboard.jsx
@@ -15,145 +15,20 @@ import Paper from '@mui/material/Paper';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 import ButtonBar from '../Components/ReleasePlan/ButtonBar';
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import { useState, useEffect } from 'react';
 import ListItemIcon from '@mui/material/ListItemIcon';
-import PersonAddIcon from '@mui/icons-material/PersonAdd';
-import EditIcon from '@mui/icons-material/Edit';
-import DeleteIcon from '@mui/icons-material/Delete';
-import SendIcon from '@mui/icons-material/Send';
-import CheckIcon from '@mui/icons-material/Check';
-import CloseIcon from '@mui/icons-material/Close';
-import { Dialog, DialogTitle, DialogContent, DialogActions, TextField } from '@mui/material'
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemText from '@mui/material/ListItemText';
-import Avatar from '@mui/material/Avatar';
-import Divider from '@mui/material/Divider';
-import NotificationsNoneIcon from '@mui/icons-material/NotificationsNone';
-import NotificationsActiveIcon from '@mui/icons-material/NotificationsActive';
-import { projectRowDataAPI, newProjectAPI, sendInviteAPI, projectUserListAPI, cancelInviteAPI, kickUserAPI, renameProjectAPI, deleteProjectAPI } from '../API/project';
-import { getUserIdAPI, acceptInviteAPI, rejectInviteAPI, fetchNotificationsAPI } from '../API/user'
+import { Dialog, DialogTitle, DialogActions } from '@mui/material'
+import { projectRowDataAPI } from '../API/project';
+import { getUserIdAPI } from '../API/user'
+import ProjectCreateButton from '../Components/ReleasePlan/ProjectCreateButton';
+import ProjectNotificationsButton from '../Components/ReleasePlan/ProjectNotificationsButton';
+import ProjectShareButton from '../Components/ReleasePlan/ProjectShareButton';
+import ProjectRenameButton from '../Components/ReleasePlan/ProjectRenameButton';
+import ProjectDeleteButton from '../Components/ReleasePlan/ProjectDeleteButton';
 
 export default function Dashboard({ setName, setSelectedProjectId }) {
 	const [rows, setRows] = useState([]);
-
-	const [createDialogOpen, setCreateDialogOpen] = useState(false);
-	const [newProjectName, setNewProjectName] = useState('');
-
-	const [invitesDialogOpen, setInvitesDialogOpen] = useState(false);
-	const [inviteList, setInviteList] = useState([]);
-
-	const [shareDialogOpen, setShareDialogOpen] = useState(false);
-	const [userList, setUserList] = useState([[], { username: "loading..." }, []]);
-	const [recipient, setRecipient] = useState('');
-	const [recipientError, setRecipientError] = useState(false);
-	const [sharingProjectId, setSharingProjectId] = useState(null);
-
-	const [renameDialogOpen, setRenameDialogOpen] = useState(false);
-	const [renameProjectTextfield, setRenameProjectTextfield] = useState('');
-	const [renameProjectId, setRenameProjectId] = useState(null);
-
-	const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
-	const [deletedProjectName, setDeletedProjectName] = useState('');
-	const [deleteProjectId, setDeleteProjectId] = useState(null);
-
 	const [ownUserId, setOwnUserId] = useState(null);
-
-	const handleCreateDialogOpen = () => {
-		setCreateDialogOpen(true);
-	};
-
-	const handleCreateDialogClose = () => {
-		setCreateDialogOpen(false);
-	};
-
-	const handleCreate = () => {
-		if (!isValidProjectName(newProjectName)) {
-			return;
-		}
-		handleCreateDialogClose();
-		fetchCreateNewProject();
-		setNewProjectName('');
-	};
-
-	const handleInvitesDialogOpen = () => {
-		setInvitesDialogOpen(true);
-	};
-
-	const handleInvitesDialogClose = () => {
-		setInvitesDialogOpen(false);
-	};
-
-	const handleShareDialogOpen = (projectId) => {
-		setSharingProjectId(projectId);
-		fetchUserList(projectId);
-		setShareDialogOpen(true);
-	};
-
-	const handleShareDialogClose = (event, reason) => {
-		setRecipientError(false)
-		if (reason !== 'backdropClick') {
-			setRecipient('');
-			setShareDialogOpen(false);
-		}
-	};
-
-	const handleShare = () => {
-		fetchSendInvite();
-	};
-
-	const handleShareEnterPress = (event) => {
-		event.preventDefault();
-		handleShare();
-	};
-
-	const confirmShare = () => {
-		handleShareDialogClose();
-		setSharingProjectId(null);
-		setRecipient('');
-	};
-
-	const handleRenameDialogOpen = (projectName, projectId) => {
-		setRenameProjectTextfield(projectName);
-		setRenameProjectId(projectId);
-		setRenameDialogOpen(true);
-	};
-
-	const handleRenameDialogClose = () => {
-		setRenameProjectTextfield('');
-		setRenameProjectId(null);
-		setRenameDialogOpen(false);
-	};
-
-	const handleRename = () => {
-		if (!isValidProjectName(renameProjectTextfield)) {
-			return;
-		}
-		fetchRenameProject();
-		handleRenameDialogClose();
-	};
-
-	const handleRenameEnterPress = (event) => {
-		event.preventDefault();
-		handleRename();
-	};
-
-	const handleDeleteDialogOpen = (name, id) => {
-		setDeletedProjectName(name);
-		setDeleteProjectId(id)
-		setDeleteDialogOpen(true);
-	};
-
-	const handleDeleteDialogClose = () => {
-		setDeleteDialogOpen(false);
-	};
-
-	const handleDelete = () => {
-		fetchDeleteProject();
-		setDeleteProjectId(null);
-		handleDeleteDialogClose();
-	};
 
 	function fetchProjectRowData() {
 		projectRowDataAPI(setRows);
@@ -163,15 +38,6 @@ export default function Dashboard({ setName, setSelectedProjectId }) {
 		fetchProjectRowData();
 	}, []);
 
-	function fetchCreateNewProject() {
-		const resultSuccessHandler = (result) => setRows(rows.concat(result));
-		newProjectAPI(newProjectName, resultSuccessHandler);
-	}
-
-	function isValidProjectName(str) {
-		return str !== "";
-	}
-
 	function fetchUserId() {
 		getUserIdAPI(setOwnUserId);
 	}
@@ -180,78 +46,12 @@ export default function Dashboard({ setName, setSelectedProjectId }) {
 		fetchUserId();
 	}, []);
 
-	function fetchUserList(projectId) {
-		const resultSuccessHandler = (response) => {
-			setUserList(response)
-		}
-		projectUserListAPI(projectId, resultSuccessHandler);
-	}
-
-	function fetchSendInvite() {
-		const resultSuccessHandler = (response) => {
-			setUserList(response);
-			setRecipient('')
-			setRecipientError(false);
-		}
-		const resultFailureHandler = (response) => {
-			setRecipientError(true)
-		}
-		sendInviteAPI(sharingProjectId, recipient, resultSuccessHandler, resultFailureHandler);
-	}
-
-	function fetchCancelInvite(userId) {
-		cancelInviteAPI(sharingProjectId, userId, setUserList);
-	}
-
-	function fetchKickTeamMember(userId) {
-		kickUserAPI(sharingProjectId, userId, setUserList);
-	}
-
-	function fetchNotifications() {
-		fetchNotificationsAPI(setInviteList);
-	}
-
-	useEffect(() => {
-		fetchNotifications();
-	}, []);
-
-	function fetchAcceptInvite(projectId) {
-		const resultSuccessHandler = (response) => {
-			setInviteList(response[0]);
-			setRows(rows.concat(response[1]));
-		}
-		acceptInviteAPI(projectId, resultSuccessHandler);
-	}
-
-	function fetchRejectInvite(projectId) {
-		rejectInviteAPI(projectId, setInviteList);
-	}
-
-	function fetchRenameProject() {
-		const resultSuccessHandler = (response) => {
-			const index = rows.findIndex(obj => obj.id === renameProjectId);
-			const rowsCopy = [...rows];
-			rowsCopy[index] = response;
-			setRows(rowsCopy);
-		}
-		renameProjectAPI(renameProjectId, renameProjectTextfield, resultSuccessHandler);
-	}
-
-	function fetchDeleteProject() {
-		const resultSuccessHandler = (response) => {
-			const filtered = rows.filter((projRowData) => { return projRowData.id !== deleteProjectId });
-			setRows(filtered);
-		}
-		deleteProjectAPI(deleteProjectId, resultSuccessHandler);
-	}
-
 	function Row(projectData) {
 		const [open, setOpen] = useState(false);
-		const data = projectData.row // may need to remove to update
+		const data = projectData.row
 		// https://forum.freecodecamp.org/t/how-to-convert-date-to-dd-mm-yyyy-in-react/431093
 		const options = { year: 'numeric', month: 'long', day: 'numeric' };
 		const formattedDate = new Date(data.dateCreated).toLocaleDateString('en-US', options);
-
 
 		return (
 			<React.Fragment>
@@ -294,27 +94,9 @@ export default function Dashboard({ setName, setSelectedProjectId }) {
 										Shortcuts
 									</Typography>
 									<ListItemIcon>
-										<IconButton
-											onClick={() => {
-												handleShareDialogOpen(data.id)
-											}}
-										>
-											<PersonAddIcon fontSize="small" />
-										</IconButton>
-										<IconButton
-											onClick={() => {
-												handleRenameDialogOpen(data.name, data.id)
-											}}
-										>
-											<EditIcon fontSize="small" />
-										</IconButton>
-										<IconButton
-											onClick={() => {
-												handleDeleteDialogOpen(data.name, data.id)
-											}}
-										>
-											<DeleteIcon fontSize="small" />
-										</IconButton>
+										<ProjectShareButton id={data.id} ownUserId={ownUserId} />
+										<ProjectRenameButton rows={rows} setRows={setRows} name={data.name} id={data.id} />
+										<ProjectDeleteButton rows={rows} setRows={setRows} name={data.name} id={data.id} />
 									</ListItemIcon>
 								</Box>
 								<Box
@@ -347,16 +129,8 @@ export default function Dashboard({ setName, setSelectedProjectId }) {
 				>
 					My Projects
 				</Typography>
-				<IconButton
-					onClick={handleCreateDialogOpen}
-				>
-					<AddCircleOutlineIcon fontSize="small" />
-				</IconButton>
-				<IconButton
-					onClick={handleInvitesDialogOpen}
-				>
-					{inviteList.length > 0 ? <NotificationsActiveIcon fontSize="small" style={{ color: 'red' }} /> : <NotificationsNoneIcon fontSize="small" />}
-				</IconButton>
+				<ProjectCreateButton rows={rows} setRows={setRows} />
+				<ProjectNotificationsButton rows={rows} setRows={setRows} />
 			</Box>
 
 			<TableContainer component={Paper} sx={{ width: '90%', margin: 'auto' }}>
@@ -398,174 +172,6 @@ export default function Dashboard({ setName, setSelectedProjectId }) {
 					</TableBody>
 				</Table>
 			</TableContainer>
-
-			<Dialog open={createDialogOpen} onClose={handleCreateDialogClose} maxWidth="sm" fullWidth>
-				<DialogTitle>Create a new project</DialogTitle>
-				<DialogContent>
-					<TextField
-						autoFocus
-						margin="dense"
-						label="Name"
-						type="text"
-						fullWidth
-						variant="outlined"
-						value={newProjectName}
-						onChange={(e) => setNewProjectName(e.target.value)}
-					/>
-				</DialogContent>
-				<DialogActions>
-					<Button onClick={handleCreateDialogClose}>Cancel</Button>
-					<Button onClick={handleCreate} color="primary">Create</Button>
-				</DialogActions>
-			</Dialog>
-
-			<Dialog open={invitesDialogOpen} onClose={handleInvitesDialogClose} maxWidth="sm" fullWidth>
-				<DialogTitle>Project invitations</DialogTitle>
-				<DialogContent>
-					<Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', flexDirection: 'column' }}>
-						<List sx={{ width: '90%', bgcolor: 'background.paper' }}>
-							{inviteList.length > 0 && <Divider />}
-							{inviteList.map(invitedProject =>
-								<>
-									<ListItem>
-										<ListItemText primary={invitedProject.name} />
-										<IconButton edge="end" aria-label="delete" onClick={() => { fetchAcceptInvite(invitedProject.id) }} sx={{ marginRight: 1 }}>
-											<CheckIcon />
-										</IconButton>
-										<IconButton edge="end" aria-label="delete" onClick={() => { fetchRejectInvite(invitedProject.id) }}>
-											<CloseIcon />
-										</IconButton>
-									</ListItem>
-									<Divider />
-								</>)}
-						</List>
-					</Box>
-				</DialogContent>
-				<DialogActions>
-					<Button onClick={handleInvitesDialogClose}>Done</Button>
-				</DialogActions>
-			</Dialog>
-
-			<Dialog
-				open={shareDialogOpen}
-				onClose={handleShareDialogClose}
-				disableEscapeKeyDown={true}
-				maxWidth="sm"
-				fullWidth
-				slotProps={{
-					backdrop: undefined
-				}}
-			>
-				<DialogTitle>Manage access to your project</DialogTitle>
-
-				<Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', flexDirection: 'column' }}>
-					<List sx={{ width: '90%', bgcolor: 'background.paper' }}>
-						<Divider />
-						{userList[0].map((user) => (
-							<>
-								<ListItem key={user.id}>
-									<ListItemIcon>
-										<Avatar>Inv</Avatar>
-									</ListItemIcon>
-									<ListItemText primary={user.username} />
-									<IconButton edge="end" aria-label="delete" onClick={() => { fetchCancelInvite(user.id) }}>
-										<CloseIcon />
-									</IconButton>
-								</ListItem>
-								<Divider />
-							</>
-						))}
-						<ListItem>
-							<ListItemIcon>
-								<Avatar>PO</Avatar>
-							</ListItemIcon>
-							<ListItemText primary={userList[1].username} />
-						</ListItem>
-						<Divider />
-						{userList[2].map((user) => (
-							<>
-								<ListItem>
-									<ListItemIcon>
-										<Avatar>TM</Avatar>
-									</ListItemIcon>
-									<ListItemText primary={user.username} />
-									{ownUserId === userList[1].id &&
-										<IconButton edge="end" aria-label="delete" onClick={() => { fetchKickTeamMember(user.id) }}>
-											<DeleteIcon />
-										</IconButton>}
-								</ListItem>
-								<Divider />
-							</>
-						))}
-					</List>
-				</Box>
-
-				<DialogContent>
-					<Box
-						component="form"
-						onSubmit={handleShareEnterPress}
-						sx={{ display: 'flex', alignItems: 'center' }}
-					>
-						<TextField
-							error={recipientError}
-							helperText={recipientError ? "User not found or is already on the list" : ""}
-							autoFocus
-							margin="dense"
-							label="Recipient email"
-							type="text"
-							variant="outlined"
-							sx={{ width: '90%', mr: 1 }}
-							value={recipient}
-							onChange={(e) => setRecipient(e.target.value)}
-						/>
-						<IconButton
-							edge="end"
-							aria-label="send"
-							onClick={handleShare}
-						>
-							<SendIcon fontSize="large" style={{ color: '#3477eb' }} />
-						</IconButton>
-					</Box>
-				</DialogContent>
-				<Box sx={{ padding: '16px 10px' }}>
-					<Button variant="outlined" onClick={confirmShare} color="primary" fullWidth>
-						Confirm List
-					</Button>
-				</Box>
-			</Dialog>
-
-			<Dialog open={renameDialogOpen} onClose={handleRenameDialogClose} maxWidth="sm" fullWidth>
-				<Box
-					component="form"
-					onSubmit={handleRenameEnterPress}
-				>
-					<DialogTitle>Rename your project</DialogTitle>
-					<DialogContent>
-						<TextField
-							autoFocus
-							margin="dense"
-							label="Name"
-							type="text"
-							fullWidth
-							variant="outlined"
-							value={renameProjectTextfield}
-							onChange={(e) => setRenameProjectTextfield(e.target.value)}
-						/>
-					</DialogContent>
-					<DialogActions>
-						<Button onClick={handleRenameDialogClose}>Cancel</Button>
-						<Button onClick={handleRename} color="primary">Rename</Button>
-					</DialogActions>
-				</Box>
-			</Dialog>
-
-			<Dialog open={deleteDialogOpen} onClose={handleDeleteDialogClose} maxWidth="sm" fullWidth>
-				<DialogTitle>Are you sure you want to leave "{deletedProjectName}"?</DialogTitle>
-				<DialogActions>
-					<Button onClick={handleDeleteDialogClose}>Cancel</Button>
-					<Button onClick={handleDelete} color="error">Leave</Button>
-				</DialogActions>
-			</Dialog>
 		</>
 	);
 }
